speedreader: declare a matching indexmap v1.9.3
When building as part of brave's speedreader component, we're
linked against v1.9.3 of speedreader, which uses a build script
to try to auto-detect whether std is available. This is new
since the 1.6.0 we originally tested with, and further, fails
on i686-linux-android targets in test builds. Fortunately we
can work around the issue by explicitly enabling the std
feature. This can be removed after upgrading to 2.x where it
is enabled by default.