impl Display for NodeRef
Replace the `std::string::ToString` impl with `std::fmt::Display`.
`ToString` is implemented in the std library for all `Display`
types, so documentation recommends implementing the more general
trait instead.

The new trait method is fallible, returning `fmt::Error` instead
of `unwrap()` on serialization failure, but the blanket `to_string`
impl converts this to a panic so the only behavior change is a
less-localized error message. Serialization shouldn't fail so
this isn't a major concern.

Addresses a clippy lint.