Fix doctests
Doctests have always been disabled in this repo. Add sufficient
annotations so that they pass and re-enable them.

As far as I can tell, the first `cell_extras` example is not
supposed to compile, demonstrating the soundness of the
extension trait. Perhaps this is why the tests were disabled.
It looks like the `compile_fail` annotation was stablized
in 2017 with rust 1.22.0.